# Make `Enumerable.pluck` faster for single key